+++ R/FI.MCN.R
@@ -26,7 +26,7 @@ X<-Xs}
   p0=p*(p+1)/2
   b=as.matrix(P[1:q])
   B=xpnd(P[(q+1):(q+p0)])
-  invB=solve(B)
+  invB=solve2(B)
   h=matrix(0, nrow=p)
   v=as.numeric(P[length(P)-1])
   g=as.numeric(P[length(P)])