+++ R/FI.MSL.R
@@ -27,7 +27,7 @@ X<-Xs}
   p0=p*(p+1)/2
   b=as.matrix(P[1:q])
   B=xpnd(P[(q+1):(q+p0)])
-  invB=solve(B)
+  invB=solve2(B)
   h=matrix(0, nrow=p)
   v=as.numeric(P[length(P)])
   d<-matrix(0,n,1)