+++ R/estimate.MSCN.R
@@ -9,7 +9,7 @@ lmscr<-function(P,y,X){
   q=ncol(X[[n]])
   b=as.matrix(P[1:q])
   B=xpnd(P[(q+1):(q+p*(p+1)/2)])
-  invB=solve(B)
+  invB=solve2(B)
   h=as.matrix(P[(q+p*(p+1)/2+1):(length(P)-2)])
   v=as.numeric(P[length(P)-1])
   g=as.numeric(P[length(P)])
@@ -24,7 +24,7 @@ lmscvgr<-function(PA,b,B,h,y,X){
   if(missing(X)) {X<-array(1,c(p,1,n))}
   v=PA[1]
   g=PA[2]
-  invB=solve(B)
+  invB=solve2(B)
   e<-matrix(0,n,p)
   for(i in 1:n){e[i,]<-y[i,]-X[[i]]%*%as.matrix(b)}
   lvg=sum(log(v/det(1/sqrt(g)*B)*mvtnorm::dmvnorm(sqrt(g)*e%*%invB)+(1-v)/det(B)*mvtnorm::dmvnorm(e%*%invB)))
@@ -37,7 +37,7 @@ escmscr<-function(P,y,X){
   b=as.matrix(P[1:q])
   p0=p*(p+1)/2
   B=xpnd(P[(q+1):(q+p0)])
-  invB=solve(B)
+  invB=solve2(B)
   h=as.matrix(P[(q+p*(p+1)/2+1):(length(P)-2)])
   v=as.numeric(P[length(P)-1])
   g=as.numeric(P[length(P)])
@@ -109,14 +109,14 @@ X<-Xs}
     b0<-b0+t(X[[i]])%*%X[[i]]
     b1<-b1+t(X[[i]])%*%y[i,]
   }
-  b<-solve(b0)%*%b1
+  b<-solve2(b0)%*%b1
   e<-matrix(0,n,p)
   for(i in 1:n){
     e[i,]<-y[i,]-X[[i]]%*%b
   }
   S<-cov(e)
   B<-matrix.sqrt(S)
-  invB<-solve(B)
+  invB<-solve2(B)
   h<-as.matrix(moments::skewness(e))
   v<-0.5
   g<-0.5
@@ -134,17 +134,17 @@ X<-Xs}
     W<-as.matrix(dnorm(aux)/pnorm(aux))
     t<-e%*%invB%*%h+W #passo E
     S<-1/n*(t(e)%*%diag(as.vector(u))%*%e)
-    invS<-solve(S)
+    invS<-solve2(S)
     B<-matrix.sqrt(S)
-    invB<-solve(B)
-    h<-B%*%solve(t(e)%*%e)%*%t(e)%*%t
+    invB<-solve2(B)
+    h<-B%*%solve2(t(e)%*%e)%*%t(e)%*%t
     b0<-matrix(0,q,q)
     b1<-matrix(0,q,1)
     for(i in 1:n){
       b0<-b0+t(X[[i]])%*%invB%*%(as.numeric(u[i])*diag(p)+h%*%t(h))%*%invB%*%X[[i]]
       b1<-b1+t(X[[i]])%*%(as.numeric(u[i])*invS%*%y[i,]-as.numeric(t[i])*invB%*%h+invB%*%h%*%t(h)%*%invB%*%y[i,])
     }
-    C<-solve(b0)
+    C<-solve2(b0)
     b<-C%*%b1
     d<-matrix(0,n,1)
     e<-matrix(0,n,p)
@@ -174,7 +174,7 @@ X<-Xs}
  if(est.var)
  {
  MI.obs<- FI.MSCN(P,y,X)
- test=try(solve(MI.obs,tol=1e-100),silent=TRUE)
+ test=try(solve2(MI.obs),silent=TRUE)
  se=c()
  if(is.numeric(test) & max(diag(test))<0) 
  {