+++ R/estimate.MSL.R
@@ -10,7 +10,7 @@ lmsr<-function(P,y,X){
   q=ncol(X[[n]])
   b=as.matrix(P[1:q])
   B=xpnd(P[(q+1):(q+p*(p+1)/2)])
-  invB=solve(B)
+  invB=solve2(B)
   v=as.numeric(P[length(P)])
   d<-matrix(0,n,1)
   e<-matrix(0,n,p)
@@ -23,7 +23,7 @@ lmsvr<-function(v,b,B,y,X){
   n=nrow(y)
   p=ncol(y)
   if(missing(X)) {X<-array(1,c(p,1,n))}
-  invB=solve(B)
+  invB=solve2(B)
   d<-matrix(0,n,1)
   e<-matrix(0,n,p)
   for(i in 1:n){
@@ -40,7 +40,7 @@ escmsr<-function(P,y,X){
   b=as.matrix(P[1:q])
   p0=p*(p+1)/2
   B=xpnd(P[(q+1):(q+p0)])
-  invB=solve(B)
+  invB=solve2(B)
   h=matrix(0, nrow=p)
   v=as.numeric(P[length(P)])
   d<-matrix(0,n,1)
@@ -106,15 +106,15 @@ X<-Xs}
     b0<-b0+t(X[[i]])%*%X[[i]]
     b1<-b1+t(X[[i]])%*%y[i,]
   }
-  b<-solve(b0)%*%b1
+  b<-solve2(b0)%*%b1
   e<-matrix(0,n,p)
   for(i in 1:n){
     e[i,]<-y[i,]-X[[i]]%*%b
   }
   S<-cov(e)
-  invS<-solve(S)
+  invS<-solve2(S)
   B<-matrix.sqrt(S)
-  invB<-solve(B)
+  invB<-solve2(B)
   h<-matrix(0, nrow=p)
   v<-1.01
   P_0<-c(as.vector(b),vech(B),as.vector(h),v)
@@ -135,16 +135,16 @@ X<-Xs}
     W<-as.matrix(dnorm(aux)/pnorm(aux))
     t<-e%*%invB%*%h+W 
     S<-1/n*(t(e)%*%diag(as.vector(u))%*%e)
-    invS<-solve(S)
+    invS<-solve2(S)
     B<-matrix.sqrt(S)
-    invB<-solve(B)
+    invB<-solve2(B)
     b0<-matrix(0,q,q)
     b1<-matrix(0,q,1)
     for(i in 1:n){
       b0<-b0+t(X[[i]])%*%invB%*%(as.numeric(u[i])*diag(p)+h%*%t(h))%*%invB%*%X[[i]]
       b1<-b1+t(X[[i]])%*%(as.numeric(u[i])*invS%*%y[i,]-as.numeric(t[i])*invB%*%h+invB%*%h%*%t(h)%*%invB%*%y[i,])  
     }
-    b<-solve(b0)%*%b1
+    b<-solve2(b0)%*%b1
     V<-optim(v,lmsvr,gr=NULL,b,B,y,X,method="L-BFGS-B",lower=nu.min,upper=100,control=list(fnscale=-1,maxit=50))   
     v<-as.numeric(V$par)
     P<-c(as.vector(b),vech(B),as.vector(h),v)
@@ -165,7 +165,7 @@ X<-Xs}
  if(est.var)
  {
  MI.obs<-FI.MSL(P,y,X)
- test=try(solve(MI.obs,tol=1e-100),silent=TRUE)
+ test=try(solve2(MI.obs),silent=TRUE)
  se=c()
  if(is.numeric(test) & max(diag(test))<0) 
  {