+++ R/estimate.MSSL.R
@@ -9,7 +9,7 @@ lmssr<-function(P,y,X){
   q=ncol(X[[n]])
   b=as.matrix(P[1:q])
   B=xpnd(P[(q+1):(q+p*(p+1)/2)])
-  invB=solve(B)
+  invB=solve2(B)
   h=as.matrix(P[(q+p*(p+1)/2+1):(length(P)-1)])
   v=as.numeric(P[length(P)])
   d<-matrix(0,n,1)
@@ -24,7 +24,7 @@ lmssvr<-function(v,b,B,h,y,X){
   n=nrow(y)
   p=ncol(y)
   if(missing(X)) {X<-array(1,c(p,1,n))}
-  invB=solve(B)
+  invB=solve2(B)
   d<-matrix(0,n,1)
   e<-matrix(0,n,p)
   for(i in 1:n){
@@ -44,7 +44,7 @@ escmssr<-function(P,y,X){
   b=as.matrix(P[1:q])
   p0=p*(p+1)/2
   B=xpnd(P[(q+1):(q+p0)])
-  invB=solve(B)
+  invB=solve2(B)
   h=as.matrix(P[(q+p0+1):(length(P)-1)])
   v=as.numeric(P[length(P)])
   d<-matrix(0,n,1)
@@ -110,15 +110,15 @@ X<-Xs}
     b0<-b0+t(X[[i]])%*%X[[i]]
     b1<-b1+t(X[[i]])%*%y[i,]
   }
-  b<-solve(b0)%*%b1
+  b<-solve2(b0)%*%b1
   e<-matrix(0,n,p)
   for(i in 1:n){
     e[i,]<-y[i,]-X[[i]]%*%b
   }
   S<-cov(e)
-  invS<-solve(S)
+  invS<-solve2(S)
   B<-matrix.sqrt(S)
-  invB<-solve(B)
+  invB<-solve2(B)
   h<-as.matrix(moments::skewness(e))
   v<-1.01
   P_0<-c(as.vector(b),vech(B),as.vector(h),v)
@@ -141,17 +141,17 @@ X<-Xs}
     W<-as.matrix(dnorm(aux)/pnorm(aux))
     t<-e%*%invB%*%h+W #passo E
     S<-1/n*(t(e)%*%diag(as.vector(u))%*%e)
-    invS<-solve(S)
+    invS<-solve2(S)
     B<-matrix.sqrt(S)
-    invB<-solve(B)
-    h<-B%*%solve(t(e)%*%e)%*%t(e)%*%t
+    invB<-solve2(B)
+    h<-B%*%solve2(t(e)%*%e)%*%t(e)%*%t
     b0<-matrix(0,q,q)
     b1<-matrix(0,q,1)
     for(i in 1:n){
       b0<-b0+t(X[[i]])%*%invB%*%(as.numeric(u[i])*diag(p)+h%*%t(h))%*%invB%*%X[[i]]
       b1<-b1+t(X[[i]])%*%(as.numeric(u[i])*invS%*%y[i,]-as.numeric(t[i])*invB%*%h+invB%*%h%*%t(h)%*%invB%*%y[i,])  
     }
-    C<-solve(b0)
+    C<-solve2(b0)
     b<-C%*%b1
     v<-max(nu.min,-n/sum(lu))
     P<-c(as.vector(b),vech(B),as.vector(h),v)
@@ -172,7 +172,7 @@ conv<-ifelse(iter<=max.iter & crit<=prec, 0, 1)
  conv.problem=1
  if(est.var){
  MI.obs<-FI.MSSL(P,y,X)
- test=try(solve(MI.obs,tol=1e-100),silent=TRUE)
+ test=try(solve2(MI.obs),silent=TRUE)
  se=c()
  if(is.numeric(test) & max(diag(test))<0) {
  conv.problem=0