+++ R/estimate.MSSLEC.R
@@ -46,7 +46,7 @@ exp(log.int)}
   m=ncol(X[[1]])
 eta<-matrix(eta,ncol=1);beta<-matrix(beta,ncol=1)
 ll<-c()
-iSigma<-solve(Sigma)
+iSigma<-solve2(Sigma)
 logdSigma<-log(det(Sigma))
 for(i in 1:n)
 {
@@ -99,7 +99,7 @@ max.int<-1
   integrate(aux.den.MSMSNC,lower=0,upper=max.int,a,d,p,dist=dist,nu=nu,abs.tol=1e-12)$value
 }
 n<-nrow(y) ; p=ncol(y)
-      iSigma<-solve(Sigma)
+      iSigma<-solve2(Sigma)
 a.theta<-c();b.theta<-c();c.theta<-c()
 for(i in 1:n)
 { 
@@ -122,13 +122,13 @@ p=ncol(X[[1]]);m=nrow(X[[1]])
 eta<-matrix(eta,ncol=1)
 sum1<-matrix(0,ncol=p,nrow=p)
 sum2<-matrix(0,nrow=p,ncol=1)
-  invSigma=solve(Sigma)
+  invSigma=solve2(Sigma)
 for(i in 1:nrow(y))
 {
 sum1<-sum1+t(X[[i]])%*%(a.theta[i]*invSigma+b.theta[i]*eta%*%t(eta))%*%X[[i]]
 sum2<-sum2+t(X[[i]])%*%(b.theta[i]*eta%*%t(eta)%*%y[i,]+a.theta[i]*invSigma%*%y[i,] - c.theta[i]*eta)   
 }
-  solve(sum1,tol=1e-100)%*%sum2
+  solve2(sum1)%*%sum2
 }
 M2.step.MSMSNC<-function(y,X,beta,a.theta)
 {
@@ -151,7 +151,7 @@ for(i in 1:n)
 sum1<-sum1+b.theta[i]*(y[i,]-X[[i]]%*%beta)%*%t(y[i,]-X[[i]]%*%beta)
 sum2<-sum2+c.theta[i]*(y[i,]-X[[i]]%*%beta)
 }
-solve(sum1,tol=1e-100)%*%sum2
+solve2(sum1)%*%sum2
 }
 n=nrow(y)
   p=ncol(y)
@@ -222,7 +222,7 @@ conv<-ifelse(i<=max.iter & dif<=prec, 0, 1)
  if(est.var)
  {
  MI.obs<-FI.MSSLEC(P,y,X,nu.new)
- test=try(solve(MI.obs,tol=1e-100),silent=TRUE)
+ test=try(solve2(MI.obs),silent=TRUE)
  if(is.numeric(test) & max(diag(test))<0) 
  {
  conv.problem=0
@@ -252,7 +252,7 @@ rownames(P)<-c(paste("beta",1:m,sep=""),paste("alpha",indices,sep=""),paste("lam
     b0<-b0+t(X[[i]])%*%X[[i]]
     b1<-b1+t(X[[i]])%*%y[i,] 
   }
-  beta0<-solve(b0)%*%b1
+  beta0<-solve2(b0)%*%b1
   e<-matrix(0,n,p)
   for(i in 1:n){
     e[i,]<-y[i,]-X[[i]]%*%beta0