+++ R/estimate.MSTT.R
@@ -5,7 +5,7 @@ logsttnu<-function(nu,Y,X,beta0,Sigma,lambda){
 n=nrow(Y)
 p=ncol(Y)
 B=matrix.sqrt(Sigma) 
-Binv=solve(B)
+Binv=solve2(B)
 mu=matrix(0,n,p)
 d<-matrix(0,n,1)
 res<-matrix(0,n,p)
@@ -14,7 +14,7 @@ for(i in 1:n){
     res[i,]<-Y[i,]-X[[i]]%*%beta0
     d[i]<-t(res[i,])%*%Binv%*%Binv%*%res[i,]
 }
-A=as.vector(t(lambda)%*%solve(B)%*%t(res))
+A=as.vector(t(lambda)%*%solve2(B)%*%t(res))
 fy=2*mvtnorm::dmvt(res, sigma = Sigma, df = nu, log = FALSE)*pt(sqrt((nu+p)/(nu+d))*A, nu+p)
 return(sum(log(fy)))}
 mstt.logL <- function(theta,Y,X){
@@ -26,7 +26,7 @@ mstt.logL <- function(theta,Y,X){
  p1=p*(p+1)/2
  B= xpnd(theta[(q0+1):(q0+p1)]) # Sigma^{1/2}
  Sigma=B%*%B
- invSigma=solve(Sigma)
+ invSigma=solve2(Sigma)
  Binv=matrix.sqrt(invSigma) # Sigma^{-1/2} = B^{-1}
  lambda=as.matrix(theta[(q0+p1+1):pth],p,1)
  nu=as.numeric(theta[pth+1])
@@ -37,7 +37,7 @@ mstt.logL <- function(theta,Y,X){
     mu[i,]<-X[[i]]%*%beta0
     res[i,]<-Y[i,]-X[[i]]%*%beta0
     d[i]<-t(res[i,])%*%Binv%*%Binv%*%res[i,]}
-A=as.vector(t(lambda)%*%solve(B)%*%t(res))
+A=as.vector(t(lambda)%*%solve2(B)%*%t(res))
 fy=2*mvtnorm::dmvt(res, sigma = Sigma, df = nu, log = FALSE)*pt(sqrt((nu+p)/(nu+d))*A, nu+p) 
 return(sum(log(fy)))
 }
@@ -77,22 +77,22 @@ if(!nu.fixed){
     b0<-b0+t(X[[i]])%*%X[[i]]
     b1<-b1+t(X[[i]])%*%y[i,] 
   }
-  beta0<-solve(b0)%*%b1
+  beta0<-solve2(b0)%*%b1
   e<-matrix(0,n,p)
   for(i in 1:n){
     e[i,]<-y[i,]-X[[i]]%*%beta0
   }
   S<-cov(e)
-  invS<-solve(S)
+  invS<-solve2(S)
   B<-matrix.sqrt(S)
   lambda<-as.matrix(moments::skewness(e))
   nu=5
   theta0<-c(as.vector(beta0),vech(B),as.vector(lambda),nu)
   log0= mstt.logL(theta0,y,X)
   Sigma=B%*%B
-  #SinvSigma=solve(Sigma)
+  #SinvSigma=solve2(Sigma)
   delta=lambda/sqrt(1+sum(as.vector(lambda)^2))
-  Delta= matrix.sqrt(solve(Sigma))%*%delta
+  Delta= matrix.sqrt(solve2(Sigma))%*%delta
   a1=Sigma-Delta%*%t(Delta)
   bb=eigen(a1)$values
   { if (sum(bb>0)==p) Gama=a1
@@ -106,19 +106,19 @@ if(!nu.fixed){
   res<-matrix(0,n,p)
   while((criterio>=prec)&&(cont<=max.iter)&&(bb>=0)){
     cont=cont+1
-    Binv=solve(B)    
+    Binv=solve2(B)    
     for (i in 1:n){
           mu[i,]<-X[[i]]%*%beta0
           res[i,]<-y[i,]-X[[i]]%*%beta0
           d[i]<-as.numeric(t(res[i,])%*%Binv%*%Binv%*%res[i,])
     }
     ### Inicio passo E
-    Gamainv=solve(Gama)
-    A=as.vector(t(lambda)%*%solve(B)%*%t(res))    
+    Gamainv=solve2(Gama)
+    A=as.vector(t(lambda)%*%solve2(B)%*%t(res))    
     u=2*gamma((nu+p+2)/2)/gamma((nu+p)/2)/(nu+d)*pt(sqrt((nu+p+2)/(nu+d))*A, nu+p+2)/pt(sqrt((nu+p)/(nu+d))*A, nu+p)
     eta1=1/pt(sqrt((nu+p)/(nu+d))*A, nu+p)/sqrt(pi)*gamma((nu+p+1)/2)/gamma((nu+p)/2)*((nu+d)^((nu+p)/2))/((nu+d+A^2)^((nu+p+1)/2))
     MT2=1/(1+as.numeric(t(Delta)%*%Gamainv%*%Delta))
-    muT=MT2*as.vector(t(Delta)%*%solve(Gama)%*%t(res))
+    muT=MT2*as.vector(t(Delta)%*%solve2(Gama)%*%t(res))
     ut = u*muT+sqrt(MT2)*eta1
     ut2=u*(muT^2)+MT2+muT*sqrt(MT2)*eta1
     ### Fim passo E
@@ -135,13 +135,13 @@ if(!nu.fixed){
         Delta1=Delta1+ut[i]*resi
         Gama0=Gama0+u[i]*resi%*%t(resi)-ut[i]*(Delta%*%t(resi)+resi%*%t(Delta))+ut2[i]*Delta%*%t(Delta)
     }
-    beta0=solve(beta01)%*%beta02
+    beta0=solve2(beta01)%*%beta02
     Delta=Delta1/sum(ut2)
     Gama=Gama0/n
     Sigma=Gama+Delta%*%t(Delta)
-    #invSigma=solve(Sigma)
+    #invSigma=solve2(Sigma)
     B=matrix.sqrt(Sigma)
-    Binv=solve(B)
+    Binv=solve2(B)
     delta=Binv%*%Delta
     lambda=delta/sqrt(1-as.numeric(t(delta)%*%delta))    
     nu=optimize(logsttnu,c(nu.min,100),y,X,beta0,Sigma,lambda,maximum=T) 
@@ -167,22 +167,22 @@ if(is.numeric(nu.fixed)){
     b0<-b0+t(X[[i]])%*%X[[i]]
     b1<-b1+t(X[[i]])%*%y[i,] 
   }
-  beta0<-solve(b0)%*%b1
+  beta0<-solve2(b0)%*%b1
   e<-matrix(0,n,p)
   for(i in 1:n){
     e[i,]<-y[i,]-X[[i]]%*%beta0
   }
   S<-cov(e)
-  invS<-solve(S)
+  invS<-solve2(S)
   B<-matrix.sqrt(S)
   lambda<-as.matrix(moments::skewness(e))
   nu=nu.fixed
   theta0<-c(as.vector(beta0),vech(B),as.vector(lambda),nu)
   log0= mstt.logL(theta0,y,X)
   Sigma=B%*%B
-  #SinvSigma=solve(Sigma)
+  #SinvSigma=solve2(Sigma)
   delta=lambda/sqrt(1+sum(as.vector(lambda)^2))
-  Delta= matrix.sqrt(solve(Sigma))%*%delta
+  Delta= matrix.sqrt(solve2(Sigma))%*%delta
   a1=Sigma-Delta%*%t(Delta)
   bb=eigen(a1)$values
   { if (sum(bb>0)==p) Gama=a1
@@ -196,19 +196,19 @@ if(is.numeric(nu.fixed)){
   res<-matrix(0,n,p)
   while((criterio>=prec)&&(cont<=max.iter)&&(bb>=0)){
     cont=cont+1
-    Binv=solve(B)    
+    Binv=solve2(B)    
     for (i in 1:n){
           mu[i,]<-X[[i]]%*%beta0
           res[i,]<-y[i,]-X[[i]]%*%beta0
           d[i]<-as.numeric(t(res[i,])%*%Binv%*%Binv%*%res[i,])
     }
     ### Inicio passo E
-    Gamainv=solve(Gama)
-    A=as.vector(t(lambda)%*%solve(B)%*%t(res))    
+    Gamainv=solve2(Gama)
+    A=as.vector(t(lambda)%*%solve2(B)%*%t(res))    
     u=2*gamma((nu+p+2)/2)/gamma((nu+p)/2)/(nu+d)*pt(sqrt((nu+p+2)/(nu+d))*A, nu+p+2)/pt(sqrt((nu+p)/(nu+d))*A, nu+p)
     eta1=1/pt(sqrt((nu+p)/(nu+d))*A, nu+p)/sqrt(pi)*gamma((nu+p+1)/2)/gamma((nu+p)/2)*((nu+d)^((nu+p)/2))/((nu+d+A^2)^((nu+p+1)/2))
     MT2=1/(1+as.numeric(t(Delta)%*%Gamainv%*%Delta))
-    muT=MT2*as.vector(t(Delta)%*%solve(Gama)%*%t(res))
+    muT=MT2*as.vector(t(Delta)%*%solve2(Gama)%*%t(res))
     ut = u*muT+sqrt(MT2)*eta1
     ut2=u*(muT^2)+MT2+muT*sqrt(MT2)*eta1
     ### Fim passo E
@@ -225,13 +225,13 @@ if(is.numeric(nu.fixed)){
         Delta1=Delta1+ut[i]*resi
         Gama0=Gama0+u[i]*resi%*%t(resi)-ut[i]*(Delta%*%t(resi)+resi%*%t(Delta))+ut2[i]*Delta%*%t(Delta)
     }
-    beta0=solve(beta01)%*%beta02
+    beta0=solve2(beta01)%*%beta02
     Delta=Delta1/sum(ut2)
     Gama=Gama0/n
     Sigma=Gama+Delta%*%t(Delta)
-    #invSigma=solve(Sigma)
+    #invSigma=solve2(Sigma)
     B=matrix.sqrt(Sigma)
-    Binv=solve(B)
+    Binv=solve2(B)
     delta=Binv%*%Delta
     lambda=delta/sqrt(1-as.numeric(t(delta)%*%delta))    
     theta=c(as.vector(beta0),vech(B),as.vector(lambda),nu)
@@ -260,7 +260,7 @@ if(is.numeric(nu.fixed)){
  if(est.var)
  {
  MI.obs<-  FI.MSTT(P,y,X,nu)
- test=try(solve(MI.obs,tol=1e-100),silent=TRUE)
+ test=try(solve2(MI.obs),silent=TRUE)
  se=c()
  if(is.numeric(test) & max(diag(test))<0) 
  {